refactor: Move webhook routes from server to module
