+++ api.planx.uk/modules/webhooks/_old/paymentRequestEvents.ts
@@ -1,11 +1,11 @@
 import { addDays } from "date-fns";
 import { Request, Response, NextFunction } from "express";
 
-import { createScheduledEvent } from "../hasura/metadata";
+import { createScheduledEvent } from "../../../hasura/metadata";
 import {
   DAYS_UNTIL_EXPIRY,
   REMINDER_DAYS_FROM_EXPIRY,
-} from "../saveAndReturn/utils";
+} from "../../../saveAndReturn/utils";
 
 /**
  * Create two "invitation" events for a payments_request record: one for the nominee and one for the agent
