+++ api.planx.uk/modules/webhooks/_old/sanitiseApplicationData/index.test.ts
@@ -1,5 +1,5 @@
 import supertest from "supertest";
-import app from "../../server";
+import app from "../../../../server";
 import * as operations from "./operations";
 
 const mockSend = jest.fn();
